// src/esp/interfaces/enums.ts
export enum ESP_COMMAND_ENUM {
    GET_WIFI_CONNECTION_STATUS = 'GET_WIFI_CONNECTION_STATUS',
    GET_AVAILABLE_NETWORKS = 'GET_AVAILABLE_NETWORKS',
    GET_POSSIBLE_NETWORKS = 'GET_POSSIBLE_NETWORKS',
    CANCEL_STREAM_LOGS = 'CANCEL_STREAM_LOGS',
    TEST_CONNECTION = 'TEST_CONNECTION',
    SEND_COMMANDS = 'SEND_COMMANDS',
    STREAM_LOGS = 'STREAM_LOGS',
    GET_LOGS = 'GET_LOGS',
    FLASH = 'FLASH',
}

export enum COMMAND_ENUM {
    GET_DEVICE_MODE = 'GET_DEVICE_MODE',
    GET_WIFI_STATUS = 'GET_WIFI_STATUS',
    RESTART_DEVICE = 'RESTART_DEVICE',
    SCAN_NETWORKS = 'SCAN_NETWORKS',
    GET_MDNS_NAME = 'GET_MDNS_NAME',
    CONNECT_WIFI = 'CONNECT_WIFI',
    SWITCH_MODE = 'SWITCH_MODE',
    SET_MDNS = 'SET_MDNS',
    SET_WIFI = 'SET_WIFI',
    PAUSE = 'PAUSE',
}